requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
pydantic-settings

# Orchestration (we'll add agents in the next batch)
langgraph
langgraph-checkpoint

# Observability
prometheus-client
prometheus-fastapi-instrumentator
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp

# Analytics
pandas
matplotlib

python-dotenv

